requirements.txt
langchain
langchain-groq
faiss-cpu
tiktoken
langchain-community
langchain-openai
chromadb
sentence-transformers
pypdf
python-dotenv
langgraph-prebuilt

#Embeddings
#langhcain-huggingface
sentence-transformers

#Langgraph
langchain
langgraph
langchain-core
langsmith


arxiv
wikipedia
langgraph-cli[inmem]
faiss-cpu
beautifulsoup4
boto3
ipykernel
unstructured
pdfminer
pymupdf
langchain_aws
langgraph_checkpoint_aws

pinecone-client==3.2.2 
langchain-pinecone
langchain-tavily
ragas
streamlit
bedrock-agentcore